유저 차단 목록 화면 screens로 분리
issue-054-userInfoPopup
